refactor InventoryCommandRepository to use JdbcBatchUtils for batch insertion of seat states
