+++ application/src/main/java/org/mandarin/booking/app/show/InventoryCommandRepository.java
@@ -2,6 +2,7 @@ package org.mandarin.booking.app.show;
 
 import java.util.List;
 import lombok.RequiredArgsConstructor;
+import org.mandarin.booking.app.JdbcBatchUtils;
 import org.mandarin.booking.domain.show.Inventory;
 import org.mandarin.booking.domain.show.SeatState.SeatStateRow;
 import org.springframework.stereotype.Repository;
@@ -12,7 +13,7 @@ import org.springframework.transaction.annotation.Transactional;
 @RequiredArgsConstructor
 class InventoryCommandRepository {
     private final InventoryRepository repository;
-    private final SeatStateBatchRepository seatStateBatchRepository;
+    private final JdbcBatchUtils jdbcBatchUtils;
 
     void insert(Inventory inventory) {
         List<SeatStateRow> rows = inventory.extractSeatStateRows();
@@ -20,6 +21,20 @@ class InventoryCommandRepository {
 
         repository.save(inventory);
 
-        seatStateBatchRepository.batchInsert(inventory.getId(), rows);
+        batchInsert(inventory.getId(), rows);
+    }
+
+    void batchInsert(Long inventoryId, List<SeatStateRow> rows) {
+        String sql = "INSERT INTO seat_state (inventory_id, seat_id, grade_id) VALUES (?, ?, ?)";
+        jdbcBatchUtils.batchUpdate(
+                sql,
+                rows,
+                (ps, row) -> {
+                    ps.setLong(1, inventoryId);
+                    ps.setLong(2, row.seatId());
+                    ps.setLong(3, row.gradeId());
+                },
+                1000
+        );
     }
 }
